+++ src/Services/Documentgenerator/DocumentgeneratorServiceBuilder.php
@@ -17,6 +17,7 @@ use Bitrix24\SDK\Attributes\ApiServiceBuilderMetadata;
 use Bitrix24\SDK\Core\Credentials\Scope;
 use Bitrix24\SDK\Services\AbstractServiceBuilder;
 use Bitrix24\SDK\Services\Documentgenerator\Document;
+use Bitrix24\SDK\Services\Documentgenerator\Numerator;
 use Bitrix24\SDK\Services\Documentgenerator\Template;
 
 #[ApiServiceBuilderMetadata(new Scope(['documentgenerator']))]
@@ -48,6 +49,23 @@ class DocumentgeneratorServiceBuilder extends AbstractServiceBuilder
             );
             $this->serviceCache[__METHOD__] = new Template\Service\Template(
                 new Template\Service\Batch($templateBatch, $this->log),
+                $this->core,
+                $this->log
+            );
+        }
+
+        return $this->serviceCache[__METHOD__];
+    }
+
+    public function numerator(): Numerator\Service\Numerator
+    {
+        if (!isset($this->serviceCache[__METHOD__])) {
+            $numeratorBatch = new Numerator\Batch(
+                $this->core,
+                $this->log
+            );
+            $this->serviceCache[__METHOD__] = new Numerator\Service\Numerator(
+                new Numerator\Service\Batch($numeratorBatch, $this->log),
                 $this->core,
                 $this->log
             );