+++ src/Services/Documentgenerator/Numerator/Result/UpdatedNumeratorBatchResult.php
@@ -1,0 +1,31 @@
+<?php
+
+/**
+ * This file is part of the bitrix24-php-sdk package.
+ *
+ * © Dmitriy Ignatenko <algonexys@gmail.com>
+ *
+ * For the full copyright and license information, please view the MIT-LICENSE.txt
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Bitrix24\SDK\Services\Documentgenerator\Numerator\Result;
+
+use Bitrix24\SDK\Core\Result\UpdatedItemBatchResult;
+
+/**
+ * Class UpdatedNumeratorBatchResult
+ *
+ * @package Bitrix24\SDK\Services\Documentgenerator\Numerator\Result
+ */
+class UpdatedNumeratorBatchResult extends UpdatedItemBatchResult
+{
+    #[\Override]
+    public function isSuccess(): bool
+    {
+        return (bool)$this->getResponseData()->getResult();
+    }
+}
+